genai_folder/sales_agent/resume_roast/requirements.txt
crewai
crewai-tools
langchain-google-genai
tavily-python
python-dotenv
PyPDF2
docx2txt